Funcionando bem, seg fault porque recebe um sinal fim de io, depois do filho ter morrido

+++ prog2.c
@@ -10,16 +10,18 @@ int main (int argc, char *argv[]){
 	int i, j;
 	printf("\nExec_time = (%d, %d, %d)",exec_time[0],exec_time[1],exec_time[2]);
 	for (i=0; i<tam; i++){
-		printf("\nPID: %d ---- Exec time: %d", getpid(), exec_time[i]);
+		//printf("\nPID: %d ---- Exec time: %d", getpid(), exec_time[i]);
 		for (j=0; j<exec_time[i]; j++){
 			printf("%d\n", getpid());
 			sleep(1);
 		}
 		//condição I/O
-		kill(getppid(),SIGUSR1);
-		sleep(3);
-		kill(getppid(),SIGCHLD);
-		//raise(SIGSTOP);
+		if(i<tam-1){
+			kill(getppid(),SIGUSR1);
+			sleep(3);
+			kill(getppid(),SIGCHLD);
+			//raise(SIGSTOP);
+		}
 	}
 	printf("\nEu, %d, Morri",getpid());
 	kill(getppid(),SIGUSR2);
